# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BizManager Purchase & Sales Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: getSuppliers
      tags: [suppliers]
      summary: List suppliers
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create supplier
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete supplier
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /suppliers/{id}/ledger:
    get:
      operationId: getSupplierLedger
      tags: [suppliers]
      summary: Get supplier ledger
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerResponse"
...
  /sales:
    get:
      operationId: getSales
      tags: [sales]
      summary: List sales
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: paymentMode
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create sale
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        gstNumber:
          type: string
        balance:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - balance
        - createdAt
    CreateSupplierRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        gstNumber:
          type: string
      required:
        - name
...
    SaleItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        sellingPrice:
          type: number
        gstPercent:
          type: number
        gstAmount:
          type: number
        totalAmount:
          type: number
      required:
        - productId
        - productName
        - quantity
        - sellingPrice
        - gstPercent
        - gstAmount
        - totalAmount
    Sale:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        totalGst:
          type: number
        grandTotal:
          type: number
        paymentMode:
          type: string
        isGstInvoice:
          type: boolean
        notes:
          type: string
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - items
        - subtotal
        - totalGst
        - grandTotal
        - paymentMode
        - isGstInvoice
        - date
        - createdAt
    CreateSaleRequest:
      type: object
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: number
              sellingPrice:
                type: number
              gstPercent:
                type: number
            required:
              - productId
              - quantity
              - sellingPrice
              - gstPercent
        paymentMode:
          type: string
          enum: [cash, credit, upi, bank]
        isGstInvoice:
          type: boolean
        notes:
          type: string
        date:
          type: string
      required:
        - items
        - paymentMode
        - isGstInvoice
...
    LedgerEntry:
      type: object
      properties:
        id:
          type: integer
        partyType:
          type: string
        partyId:
          type: integer
        type:
          type: string
        description:
          type: string
        debit:
          type: number
        credit:
          type: number
        balance:
          type: number
        referenceId:
          type: integer
        referenceType:
          type: string
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - partyType
        - partyId
        - type
        - description
        - debit
        - credit
        - balance
        - date
        - createdAt
    LedgerResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEntry"
        openingBalance:
          type: number
        closingBalance:
          type: number
        totalDebit:
          type: number
        totalCredit:
          type: number
      required:
        - entries
        - openingBalance
        - closingBalance
        - totalDebit
        - totalCredit
...